Skip chained errors if suppressed
In PEP-415, the __suppress_context__ attribute was added to
BaseException, when set to true the __context__ of the exception should
not be printed.  This change adds support for this option, so that the
exception in rollbar will look more similar to how they would look when
printed locally.

This attribute was added in Python 3.3, I have stuck with the
convention in this function of using getattr, but I suspect that it is
no longer needed with supported python versions.